获取最新代码

+++ scripts/gulp/icons/index.ts
@@ -1,7 +1,7 @@
-import { copySync, ensureDirSync, readdirSync, readFileSync, removeSync, writeFileSync } from 'fs-extra'
+import { copySync, ensureDirSync, existsSync, readdirSync, readFileSync, writeFile, writeFileSync } from 'fs-extra'
 import { camelCase, upperFirst } from 'lodash'
 import { join } from 'path'
-import SVGO from 'svgo'
+import { optimize, OptimizeOptions } from 'svgo'
 import { buildConfig } from '../buildConfig'
 
 const definitionTemplate = `export const {{definitionName}} = {
@@ -9,33 +9,35 @@ const definitionTemplate = `export const {{definitionName}} = {
   svgString: '{{svgString}}',
 }
 `
-const { iconAssetsDir, iconDefinitionsDir, siteIconAssetsDir } = buildConfig
+const { iconAssetsDir, iconDefinitionsDir } = buildConfig
+const { iconAssetsDir: siteIconAssetsDir } = buildConfig.site
 const outputDefinitionNames: string[] = []
 
-const options: SVGO.Options = {
-  plugins: [{ removeAttrs: { attrs: ['fill', 'class'] } }, { sortAttrs: true }, { removeDimensions: true }],
+const options: OptimizeOptions = {
+  plugins: [
+    { name: 'removeAttrs', params: { attrs: ['fill', 'class'] } },
+    { name: 'sortAttrs' },
+    { name: 'removeDimensions' },
+  ],
 }
-
-const svgo = new SVGO(options)
 
 export async function generateIcons(): Promise<void> {
   const iconDirname = join(iconAssetsDir)
   const iconPaths = readdirSync(iconDirname)
 
-  const outputIcons = iconPaths.map(async iconName => {
+  iconPaths.forEach(async iconName => {
     const iconFile = join(iconAssetsDir, iconName)
     const iconFileContent = readFileSync(iconFile, 'utf8')
-    await output(iconFileContent, iconName)
+    output(iconFileContent, iconName)
   })
-  await Promise.all(outputIcons)
 
   const indexContent =
     outputDefinitionNames.map(item => `export { ${upperFirst(item)} } from './${item}'`).join('\n') + '\n'
-  writeFileSync(join(iconDefinitionsDir, `index.ts`), indexContent, 'utf8')
+  writeFile(join(iconDefinitionsDir, `index.ts`), indexContent, 'utf8')
 }
 
 async function output(content: string, iconName: string) {
-  const { data } = await svgo.optimize(content)
+  const { data } = optimize(content, options)
   outputIcons(iconName, data)
   outputDefinitions(iconName, data)
 }
@@ -59,6 +61,8 @@ function outputDefinitions(iconName: string, data: string) {
 }
 
 export function copyToSite(): void {
-  removeSync(siteIconAssetsDir)
-  copySync(iconAssetsDir, siteIconAssetsDir)
+  // 不存在的时候才 copy
+  if (!existsSync(siteIconAssetsDir)) {
+    copySync(iconAssetsDir, siteIconAssetsDir)
+  }
 }
